+++ infrastructure/app/db-secrets.tf
@@ -1,8 +1,6 @@
 # Data sources for database secrets created by the db module
-data "aws_secretsmanager_secret" "stage_db_password" {
-  name = "spellbot/stage/db-details"
-}
+data "aws_secretsmanager_secret" "db_password" {
+  for_each = local.env_names
 
-data "aws_secretsmanager_secret" "prod_db_password" {
-  name = "spellbot/prod/db-details"
+  name = "spellbot/${each.key}/db-details"
 }